Improve time calculation in clock script
This commit changes the way how time is calculated in the clock script. It fixes the calculation of total_minutes to correctly handle negative inputs for hours and minutes. It defines how to format negative time as time before midnight. Also, it includes a newline character in the output, which improves readability.