Update argument validation in clock.sh
Updated the validate_parameters function in the clock.sh script to improve error signaling. Now an error is thrown only when the number of arguments is neither 2 nor 4. This refactoring offers a cleaner and more efficient way of handling invalid arguments and confirming if the input hours and minutes are indeed integers.