Simplify and correct nursery_rnnyme in house.sh
Refactoring in the script to shorten the 'verb' array syntax. For the function nursery_rhyme, changed verse's value assignment to straightforward mathematical subtraction for increased readability and reduced the chance of miscalculations. For the main function, renamed and simplified 'from' and 'to' variables to 'verse' and 'end_verse'. Also, replaced the for loop with a while loop, utilizing a post-increment to iterate through verses. These changes make the code more readable and accurate.