Refactor README.md for java exercises and go exercises
Revised README documents in java and go exercises to improve clarity and readability. Initialized with a general introduction to the exercise and added a more comprehensive instructions section. Removed redundant and/or irrelevant information to streamline the content. Added a new structure to READMEs highlighting problem description, examples, and contributors. Also updated the markdown formatting for a cleaner presentation.