Add functionality to parse Markdown syntax into HTML
Added classes and tests in order to parse given strings with Markdown syntax. The classes return the associated HTML for that string. The pre-existing tests ensure that this code works as expected. The primary focus was to rewrite the code to optimize readability and maintainability while ensuring all tests pass.