Improve string concatenation performance in Markdown class
Replaced the use of String with StringBuilder for concatenating the result in the `parse` method of the `Markdown` class. This change is made to improve the performance, as StringBuilder is more efficient than String for concatenating strings in a loop. This will greatly reduce the time complexity of the method when dealing with large inputs.