Introduced Parser interface and refactored classes to improve code organization.
Introduced a Parser interface to make the code more cohesive and maintainable. All individual parser classes are now implementing the Parser interface, promoting code reuse and better type safety. Moreover, the LineParserFabric class was introduced to centralize the decision of which parser to activate based on the line content, making the MarkdownParser class cleaner and easier to understand. The flexible structure of the apply method in the parser classes had been maintained to implement additional rules in the future.