Improve readability in MarkdownParser class
Refactored the conditional statement in 'shouldStartList' method to use 'Predicate.not'. The 'not()' static import from 'java.util.function.Predicate' was used to make the code more outstanding and simpler to read. The modified line with 'and()' and 'not()' functions is arguably more readable as it displays the logic flow clearly.