Rename LineParser class and refactor MarkdownParser
Changed the name of the class from 'LineParserFabric' to 'LineParserFactory' to better reflect its functionality. In addition, refactored the code in 'MarkdownParser' class for better readability. The conditional statement in 'shouldStartList' method has been streamlined by combining predicates using 'and' and 'not', providing clearer logic flow. The Header Parser is also now a constant in the Parser interface. Removed unnecessary comments from the 'HeaderParserBuilder'.