Change parameter name in evaluateProgram
Parameter name in the "evaluateProgram" method was changed from "strings" to "commands". This change was implemented to make the code easier to read and make it clearer for other developers what the function is expecting as input. The parameter "commands" provides clearer context on what values should be passed into the method.

Signed-off-by: Jegors Čemisovs <jegors_cemisovs@epam.com>