Add 'Wordy' exercise
Introduced a new exercise 'Wordy' to aid learners in understanding mathematical operations like addition, subtraction, multiplication, and division, and translation of wordy problems to numerical expressions.

The README.md file was added with the instructions and objectives for the exercise. Test files, wordy_test.go and cases_test.go, were added to monitor solution progress and troubleshoot errors. The HELP.md file was added for additional guidance in case of blockers and to provide submission instructions.

The go.mod file was updated to keep track of the dependencies of the given exercise module and to guarantee its reproducible build.