Enable subtraction operation tests in ForthEvaluator
Unignored tests '@Ignore("Remove to run test")' related to subtraction operation in the ForthEvaluatorTest class. Subtraction functionality has been added within the evaluateOperation method in the ForthEvaluator class, now throwing an IllegalArgumentException when there are fewer than 2 values in the stack for subtraction. These changes enable the ForthEvaluator to correctly implement and test the subtraction operation.