Unignore tests and add implementation for operations in ForthEvaluator
Enabled previously ignored tests associated with multiplication, division, and duplication in ForthEvaluatorTest.java.

In ForthEvaluator.java, the implementation for addition, subtraction, multiplication, and division has been updated to provide more explicit error messages. The datatype for "stack" is also changed to Deque from Queue to facilitate the duplication operation. The new "requireStackHasAtLeast" method ensures there are enough values in the stack for each operation, improving the robustness of our code.