Implement 'drop' operation and related tests in ForthEvaluator
Unignored tests for 'drop' operation in ForthEvaluatorTest.java as 'drop' operation is now implemented.

In ForthEvaluator.java, implemented 'drop' operation. It removes the last (top) value on the stack. Implemented 'requireStackHasAtLeast' method that throws an error if there are not enough values on the stack for a specific operation, which ensures the robustness of the 'drop' operation code.