Implemented 'over' operation and enabled corresponding tests in ForthEvaluator
The 'over' operation has been added in ForthEvaluator.java. It copies the second to top element on the stack to the top. To ensure the stack has at least 2 values for this operation, error handling is included by using the 'requireStackHasAtLeast' method.

Also, in ForthEvaluatorTest.java, the 'Remove to run test' annotations are taken out to enable the unit tests for the 'over' operation.