Enable running previously ignored tests for ForthEvaluator
Removed 'Ignore' annotations from certain tests in ForthEvaluatorTest.java. These tests were previously being skipped, but should now be executed as changes in the ForthEvaluator.java file have made them relevant.

Changes in ForthEvaluator.java included addition of a 'macros' map and the refactoring of the evaluateLine method. This method now checks if a macro should be evaluated and calls the appropriate method. The changes allow user-defined operators to consist of built-in operators and be evaluated in the correct order, to redefine a user-defined operator, override built-in word and arithmetic operators.