Remove ForthShell, introduce ForthEngine for Forth project
ForthShell interface was removed due to lack of necessity and the ForthEngine interface was introduced instead. The getStackAsList() method from ForthShell was moved to ForthStack and renamed to asList(). The ForthCore class was modified to accommodate these changes. This refactoring enhances cohesion among classes involved in interpreting and running the Forth programs.