Add logic to 'verse' method in FoodChain.java
Added the logic to the 'verse' method in FoodChain class to return the string verse form given an integer input. Introduced two array constants ANIMALS and VERSES to store the animals and the verses of the FoodChain song respectively. This enables us to pull the correct verse based on the input integer. The 'verses' method remains to be developed.