Add test cases and helper classes for Alphametics problem
This commit introduces test cases for the Alphametics problem which is a type of cryptarithm in which a set of words is written down in the form of an equation, each letter representing a digit. The test cases cover different situations, like testing three letters, unique value, leading zero, two digits final carry etc.

Also a custom exception namely UnsolvablePuzzleException is introduced to be thrown for the situation where the problem is unsolvable.

Finally, a README.md, HELP.md and build.gradle file have been added to further assist the user.