Commit message:
Replace ForthStack interface with Java Deque utility

Deleted ForthStack interface and changed all references to it in every class to use Java's Deque utility instead. This was done for enhancing the program's robustness by exploiting Java's in-built utilities and reducing the custom code that had to be maintained. With this change, the stack operations are now more standard and aligned with Java's stack handling conventions.