Add exercise Pascal's Triangle
This commit adds a new exercise, Pascal's Triangle, to the Java track. The exercise aims to help learners understand and implement the logic of Pascal's triangle. It comes with comprehensive instructions, stubs for code and tests, and documentation to help learners get started and understand the problem. Additionally, the gradle build files have been added to handle dependencies and facilitate testing. A helpful README is also included to give learners more context and instructions for the exercise. Please note that this exercise is of difficulty 5, and hence doesn't come with any starter implementation.