Add ISBN Validator exercise
This commit adds the exercise "ISBN Verifier" to the BASH track. The logic of this verifier is to validate book identification numbers which are ISBN-10 format and consists of 9 digits plus a check character. The validation process follows a formula provided in the README.md file. Several testing cases (isbn_verifier.bats) include handling valid and invalid ISBN numbers, which are either with or without separating dashes.