Add initial implementation of Forth exercise
This commit includes the initial version of Forth exercise on the Bash track of Exercism. It provides README.md for exercise instructions, HELP.md for basic help about the exercise, forth.sh as the main script file, and forth.bats for tests. New words/definitions can be created in this exercise. Currently, support for only signed integers of at least 16 bits size is provided.