Add Tournament.ps1 script, test cases, README.md and HELP.md
This commit introduces a new script - Tournament.ps1, which creates a tally table according to the results of a football competition. It processes input strings that represent matches and their outcomes and outputs a sorted table with stats for each participating team. This update is necessary because it calculates and represents the result of the tournament in a clear and insightful way.

Also, README.md and HELP.md files were added for informative purposes. Besides that, the corresponding .ps1 test case file was added to ensure the functionality of the script.